* commit '2e946870e2238553be39853c6e795ebb2020f641':
  Fix some broken debug statements and asserts.

+++ lib/Renderscript/RSInfoExtractor.cpp
@@ -362,7 +362,7 @@ RSInfo* RSInfo::ExtractFromSource(const Source& pSource, const DependencyHashTy&
   result->mHeader.hasDebugInformation =
       static_cast<uint8_t>(module.getNamedMetadata("llvm.dbg.cu") != nullptr);
 
-  assert((cur_string_pool_offset == string_pool_size) &&
+  assert(((size_t)cur_string_pool_offset == string_pool_size) &&
             "Unexpected string pool size!");
 
   return result;
